- disable tenant data removal by default - remove handler for deleted capset - restrict deletion of capability / capset if it's assigned to loadable permission

+++ src/main/java/org/folio/roles/service/CustomTenantService.java
@@ -10,6 +10,7 @@ import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.liquibase.FolioSpringLiquibase;
 import org.folio.spring.service.TenantService;
 import org.folio.tenant.domain.dto.TenantAttributes;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
@@ -21,13 +22,16 @@ public class CustomTenantService extends TenantService {
 
   private final KafkaAdminService kafkaAdminService;
   private final List<ReferenceDataLoader> referenceDataLoaders;
+  private final boolean forcePurge;
 
   public CustomTenantService(JdbcTemplate jdbcTemplate, FolioExecutionContext context,
-                             FolioSpringLiquibase folioSpringLiquibase, KafkaAdminService kafkaAdminService,
-                             List<ReferenceDataLoader> referenceDataLoaders) {
+    FolioSpringLiquibase folioSpringLiquibase, KafkaAdminService kafkaAdminService,
+    List<ReferenceDataLoader> referenceDataLoaders,
+    @Value("${application.tenant-service.force-purge:false}") boolean forcePurge) {
     super(jdbcTemplate, context, folioSpringLiquibase);
     this.kafkaAdminService = kafkaAdminService;
     this.referenceDataLoaders = referenceDataLoaders;
+    this.forcePurge = forcePurge;
   }
 
   @Override
@@ -50,10 +54,13 @@ public class CustomTenantService extends TenantService {
 
   @Override
   public void deleteTenant(TenantAttributes tenantAttributes) {
-    /*if (tenantExists()) {
-      log.warn("Tenant's data cannot be deleted to avoid inconsistency between the module and Keycloak. "
-        + "Please manually drop the schema if it's really necessary: schema = {}", getSchemaName());
-    }*/
-    super.deleteTenant(tenantAttributes);
+    if (tenantExists()) {
+      if (forcePurge) {
+        super.deleteTenant(tenantAttributes);
+      } else {
+        log.warn("Tenant's data cannot be deleted to avoid inconsistency between the module and Keycloak. "
+          + "Please manually drop the schema if it's really necessary: schema = {}", getSchemaName());
+      }
+    }
   }
 }
